+++ client/src/scripts/uim/aitems.ts
@@ -49,8 +49,8 @@ export class AItemsModule extends UIModule<Game> {
             <span class="count"></span>
         `
 
-        el.dataset.drop_kind = "2"
-        el.dataset.drop = def.idNumber!.toString()
+        el.dataset.item_type = "2"
+        el.dataset.item_value = def.idNumber!.toString()
 
         el.addEventListener("mousedown", this.game.ui.handle_slot_click.bind(this.game.ui))
         el.addEventListener("touchstart", this.game.ui.handle_slot_touch.bind(this.game.ui))